+++ FrontEnd/rebon-frontend/src/pages/Mypage/Edit/index.js
@@ -95,12 +95,12 @@ export default function Edit() {
             setUserAgreed(response.data.agreed);
           })
           .catch((error) => {
-            console.log(error);
+            alert(error.response.data.message);
           });
       })
       .catch(function (error) {
-        console.log(error);
-        console.log(error.message);
+        // console.log(error);
+        alert(error.response.data.message);
         // 오류발생시 실행 -- 이 400일 경우, alert error 출력, 닉네임 input 공백,
         // setName(userInfo);
         // setAlertState({ display: 'block', check: 'error', message: '이미 있는 아이디 입니다' });
